requirements.txt
addict==2.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
einops==0.8.0
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
jinja2==3.1.4
kiwisolver==1.4.7
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mat4py==0.6.0
matplotlib==3.7.5
mdurl==0.1.2
mmengine==0.10.5
mmpretrain==1.2.0
model-index==0.1.11
modelindex==0.0.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-python==4.10.0.84
ordered-set==4.1.0
packaging==24.2
pillow==10.4.0
platformdirs==4.3.6
psutil==6.1.0
pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
six==1.16.0
sympy==1.13.3
termcolor==2.4.0
tomli==2.2.1
tqdm==4.67.1
triton==3.0.0
typing-extensions==4.12.2
urllib3==2.2.3
yapf==0.43.0
zipp==3.20.2
